Depend on ING Corda fork on Github Packages
UNfortunately, at this time, Github Packages does not properly support
SNAPSHOTs, so we have to bump the version of our corda fork every time
we change it. This is not so bad, because it is not expected to happen
that often and also we should depend on a specific version anyway.