Moving code style check to compilation phase
In our project, we use Spotless to autoformat and check our code style.
We decided to treat code style violations that remain after
autoformatting similar to syntax errors. The reason is that since it
already was configured to break the build, we wanted to prevent that
developers only develop their code style violations late in the process,
when they do a full build. Since we consider it breaking, we fail early.

The same is NOT true for detekt. Detekt is not about style, but
about code smells. This is a far softer type of check, and can be done
later in the process. Especially because in certain cases it is
acceptable to suppress the issue.

Both of these choices will be continuously re-evaluated. If it doesn't
work as we expect, we may change it again.