Extract ZKNetworkParameters from metadata
This commit introduces separate ZKNetworkParameters and a ServiceLoader interface to register a provider.

ZKNetworkParameters can be set in two ways:

- Through the existence of a ZKNetworkParametersProvider implementation on the class path. This is the way our users will set it. Downside of this is that we can have only one of them, making testing multiple parameters difficult.
- Through constructor injection into the ZKTransactionBuilder. This can be used for testing multiple parameters. Downside is that we can't use this in our replacements of core flows such as ZKFinalityFlow. This is because we want to keep the API as similar as possible to standard Corda, so we don't add extra constructor params.

Because of this, the flow tests of the protocol module must be moved to the integration-tests module: they make use of the core flow replacements and therefore require a ZKNetworkParametersProvider to be registered. The KSP compiler plugin can't be applied toprotocol unfortunately, because it creates a circular dependency.

Fixes #62