Set ksp.incremental: false
When KSP uses incremental compilation, we observe the following
behaviour on the ImplementationsSymbolProcessor, which writes meta-inf
files for implementations of an interface

- one implementation on the class path, clean build: meta-inf file
  written correctly
- add another impl, normal,  incremental build: only the newly added
  implementation is seen as 'dirty' and is used to create the meta-inf
  file. Result is that the meta-inf file no longer contains the first
  implementation, but only the second.

We should investigate if and how we can fix this behaviour in the KSP
plugin code, but until then, we can address this by setting KSP to be
non-incremental. Of course, the downside is that this makes it slower.